+++ frontend/src/features/statistics/components/CategoryStatistics.vue
@@ -1,8 +1,7 @@
 <script setup>
 import {ref, onMounted, nextTick} from 'vue'
-import axios from 'axios'
 import {Chart} from 'chart.js/auto'
-import api from "@/lib/api.js";
+import api from "@/lib/axios.js";
 
 const chartRef = ref(null)
 let chartInstance = null
@@ -74,7 +73,6 @@ const drawChart = () => {
 
 const loadData = async () => {
   try {
-    // const res = await axios.get(`${import.meta.env.VITE_API_BASE_URL}/statistics/category`)
     const res = await api.get(`${import.meta.env.VITE_API_BASE_URL}/statistics/category`)
     const list = res.data?.data?.categoryRates || []
     labels.value = list.map(item => item.categoryName)
