+++ frontend/src/features/statistics/components/MonthlyStatistics.vue
@@ -1,9 +1,8 @@
 <script setup>
 import {ref, watch, nextTick, computed} from 'vue'
-import axios from 'axios'
 import {Chart} from 'chart.js'
 import {format} from 'date-fns'
-import api from "@/lib/api.js";
+import api from "@/lib/axios.js";
 
 const props = defineProps({date: Date})
 const chartRef = ref(null)
@@ -50,7 +49,6 @@ const drawChart = (rate) => {
 
 const loadData = async () => {
   const dateStr = format(props.date, 'yyyy-MM-dd')
-  // const res = await axios.get(`${import.meta.env.VITE_API_BASE_URL}/statistics/monthly`, {params: {date: dateStr}})
   const res = await api.get(`${import.meta.env.VITE_API_BASE_URL}/statistics/monthly`, {params: {date: dateStr}})
   const rate = res.data.data.achievementRate
   await nextTick()
