+++ frontend/src/lib/user.js
@@ -1,19 +1,22 @@
-/* user 관련 api 호출 */
-import axios from "axios";
-
-
+import api from './axios.js';
+import { showErrorToast } from '@/utill/toast.js';
 
 export function refreshUserToken() {
-    const refreshToken = localStorage.getItem('refreshToken');
+  const refreshToken = localStorage.getItem('refreshToken');
 
-    return axios.post(
-        `${import.meta.env.VITE_API_BASE_URL}/auth/refresh`,
-        { refreshToken }, // body
-        {
-            headers: {
-                Authorization: `Bearer ${refreshToken}` // 헤더
-            },
-            withCredentials: true
-        }
-    );
+  // 리프레시 토큰 보유 여부에 따른 에러 추가
+  if (!refreshToken) {
+    showErrorToast('리프레시 토큰이 없습니다. 다시 로그인해주세요.');
+    return Promise.reject('No refresh token found');
+  }
+
+  return api.post(
+    '/auth/refresh',
+    { refreshToken },
+    {
+      headers: {
+        Authorization: `Bearer ${refreshToken}`
+      }
+    }
+  );
 }
